perf(frontend): extraction CoiffureCard memo + helpers ClientHomePage (I11)
ClientHomePage.tsx etait monolithique (1522 lignes, 17 useState) :
chaque frappe dans la barre de recherche OU dans le formulaire de
reservation provoquait un re-render des 8 cartes du catalogue,
meme si leurs props n avaient pas change. Visible sur mobile /
3G en pic Tabaski.

Refactor focalise sur le hotspot principal :

- client.helpers.ts : extraction des helpers de formatage (currency,
  duration, dates), de calcul (discount, deposit) et de logique de
  jours de fermeture. Necessaire pour permettre le memo des
  sous-composants extraits sans casser les imports.

- components/RatingStars.tsx : extrait + React.memo. Utilise dans la
  liste des avis (potentiellement >10 occurrences dans le modal).

- components/CoiffureCard.tsx : extrait + React.memo. C est la carte
  la plus rendue (8 instances dans le grid catalogue).

- Pour que React.memo bite reellement, callbacks stabilises avec
  useCallback (toggleFavorite, openDetails). Sans ca, chaque render
  cree de nouveaux callbacks -> props du child changent -> memo casse.

- isFavorite passe en booleen pre-calcule plutot que de passer le
  tableau favoriteIds entier (sinon toute modif d un favori
  re-renderait toutes les cartes).

ClientHomePage passe de 1522 a ~1320 lignes.

Reporte (hors scope I11) : extraction du modal de reservation et des
formulaires booking/avis. La logique de state y est lourdement
interconnectee (form state - availability - settings), une refonte
necessite plus de temps que ne l autorise un audit pre-prod. Le
refactor actuel resout deja le re-render le plus visible cote
utilisateur final (typing dans la search bar).
